Refactor to use interfaces instead of GameAnalyzer
Rename interfaces to be Scorer and StatusChecker, to be more descriptive of
role. Refactor classes to call interfaces instead of GameAnalyzer. Also
inject GameAnalyzer and MoveHistory dependency for ComputerPlayer
into constructor, which allows GamePlayer interface to have just a single
move() method to be overridden in classes. Rework GameRecord to contain player
order information instead of GameAnalyzer and refactor accordingly.