Split up GameAnalyzer into two classes
Split up GameAnalyzer into StandardRules and MinimaxScorer. StandardRules
implements StatusChecker and MinimaxScorer implements Scorer. MinimaxScorer
takes a StatusChecker parameter in constructor so that it can check for win/tie
states. Reorganize file tree so that StandardRules is in rules/ directory and
MinimaxScorer is in scorer/ directory. GameAnalyzer (not yet removed) is in
rules/. Rename all references accordingly and split up GameAnalyzerTest into two
test classes. Update all references to GameAnalyzer to point to either a Scorer or a
StatusChecker. Remove GameAnalyzerTest class as well.
Also add methods for passing on Board info to MoveHistory.